+++ crates/core/src/utils/archive.rs
@@ -1,9 +1,12 @@
 use super::process::ProtoProcessError;
 use proto_pdk_api::ArchiveSource;
+use rustc_hash::FxHashMap;
 use starbase_archive::{ArchiveError, Archiver};
+use starbase_styles::{Style, Stylize};
 use starbase_utils::fs::FsError;
+use starbase_utils::glob::{GlobError, GlobWalkOptions};
 use starbase_utils::net::{DownloadOptions, NetError};
-use starbase_utils::{fs, net};
+use starbase_utils::{fs, glob, net};
 use std::path::{Path, PathBuf};
 use thiserror::Error;
 use warpgate::extract_file_name_from_url;
@@ -20,11 +23,22 @@ pub enum ProtoArchiveError {
 
     #[diagnostic(transparent)]
     #[error(transparent)]
+    Glob(#[from] Box<GlobError>),
+
+    #[diagnostic(transparent)]
+    #[error(transparent)]
     Net(#[from] Box<NetError>),
 
     #[diagnostic(transparent)]
     #[error(transparent)]
     Process(#[from] Box<ProtoProcessError>),
+
+    #[diagnostic(code(proto::archive::invalid_rewrite))]
+    #[error(
+        "Invalid archive prefix rewrite {}. Paths must be relative and cannot traverse outside the archive.",
+        .path.style(Style::File),
+    )]
+    InvalidRewritePath { path: String },
 }
 
 impl From<ArchiveError> for ProtoArchiveError {
@@ -36,6 +50,12 @@ impl From<ArchiveError> for ProtoArchiveError {
 impl From<FsError> for ProtoArchiveError {
     fn from(e: FsError) -> ProtoArchiveError {
         ProtoArchiveError::Fs(Box::new(e))
+    }
+}
+
+impl From<GlobError> for ProtoArchiveError {
+    fn from(e: GlobError) -> ProtoArchiveError {
+        ProtoArchiveError::Glob(Box::new(e))
     }
 }
 
@@ -94,7 +114,7 @@ pub async fn download_and_unpack(
     if should_unpack(src, target_dir)? {
         let archive_file = download(src, temp_dir, options).await?;
 
-        unpack_source(src, target_dir, temp_dir, &archive_file).await?;
+        unpack_source(src, target_dir, &archive_file).await?;
     }
 
     Ok(())
@@ -103,10 +123,9 @@ pub async fn download_and_unpack(
 pub async fn unpack_source(
     src: &ArchiveSource,
     target_dir: &Path,
-    temp_dir: &Path,
     archive_file: &Path,
 ) -> Result<(String, PathBuf), ProtoArchiveError> {
-    let result = unpack(target_dir, temp_dir, archive_file, src.prefix.as_deref()).await;
+    let result = unpack(target_dir, archive_file, src.prefix.as_deref()).await;
 
     fs::write_file(target_dir.join(".archive-url"), &src.url)?;
 
@@ -115,7 +134,6 @@ pub async fn unpack_source(
 
 pub async fn unpack(
     target_dir: &Path,
-    _temp_dir: &Path,
     archive_file: &Path,
     prefix: Option<&str>,
 ) -> Result<(String, PathBuf), ProtoArchiveError> {
@@ -126,4 +144,50 @@ pub async fn unpack(
     }
 
     Ok(archiver.unpack_from_ext()?)
+}
+
+pub fn move_and_rewrite(
+    source_dir: &Path,
+    target_dir: &Path,
+    rewrites: &FxHashMap<String, String>,
+) -> Result<(), ProtoArchiveError> {
+    fn is_invalid(value: &str) -> bool {
+        value.starts_with("/")
+            || value.starts_with("\\")
+            || value.contains("../")
+            || value == ".."
+            || Path::new(value).is_absolute()
+    }
+
+    // Rewrite all directories from the source directory first
+    for (from_glob, to_prefix) in rewrites {
+        if is_invalid(&from_glob) || is_invalid(&to_prefix) {
+            return Err(ProtoArchiveError::InvalidRewritePath {
+                path: format!("{} -> {}", from_glob, to_prefix),
+            });
+        }
+
+        for dir in glob::walk_fast_with_options(
+            source_dir,
+            [from_glob],
+            GlobWalkOptions {
+                only_dirs: true,
+                ..Default::default()
+            },
+        )? {
+            fs::rename(dir, target_dir.join(to_prefix))?;
+        }
+    }
+
+    // Then move all remaining files last
+    for entry in fs::read_dir(source_dir)? {
+        let path = entry.path();
+
+        fs::rename(
+            &path,
+            target_dir.join(path.strip_prefix(source_dir).unwrap()),
+        )?;
+    }
+
+    Ok(())
 }